FEATURE/MEDIUM: add support for additional http-check directives introduced in HAProxy 2.2
- http-check comment
- http-check connect
- http-check disable-on-404
- http-check expect
- http-check send
- http-check send-state
- http-check set-var
- http-check unset-var

This is a re-implementation of the existing http-check parser to instead use the HTTP actions style, similar to `http-request` and `http-response`. `set-var` and `unset-var` re-use the existing action implementations.

This is a backwards incompatible change, because the existing http-check parser is replaced with this, and both can't be registered at the same time.